+++ commands/music/stop.js
@@ -8,10 +8,10 @@ module.exports = {
 		const serverQueue = queue.get(message.guild.id);
 		if (!message.member.voice.channel) return message.channel.send("You're not in a vocal channel");
 		if (!serverQueue) return message.channel.send('Nothing is playing in this server!');
-		if (serverQueue.songs[0].author.id !== message.author.id) return message.channel.send(new Discord.MessageEmbed() .setColor('RANDOM') .setDescription(`Only **${serverQueue.songs[0].author.username}** can stops the queue, beacause he requested the current song`))
+		if (serverQueue.songs[0].author.id !== message.author.id) return message.channel.send(new Discord.MessageEmbed() .setAuthor(client.user.tag, client.user.avatarURL({format: 'png', dynamic: true, size: 2048})) .setColor('RANDOM') .setDescription(`Only **${serverQueue.songs[0].author.username}** can stops the queue, beacause he requested the current song`))
 		serverQueue.songs = [];
 		serverQueue.message.delete();
 		serverQueue.connection.dispatcher.end();
-		message.channel.send(new Discord.MessageEmbed() .setColor('RANDOM') .setDescription('Music sucessfully stopped! ') .setFooter('Made by Lumap#0149')).then(m => {setTimeout(() => {m.delete()}, 15000)})
+		message.channel.send(new Discord.MessageEmbed().setAuthor(client.user.tag, client.user.avatarURL({format: 'png', dynamic: true, size: 2048})) .setColor('RANDOM') .setDescription('Music sucessfully stopped! ') .setFooter('Made by Lumap#0149')).then(m => {setTimeout(() => {m.delete()}, 15000)})
 	},
 };
